--- webapp/src/main/java/com/involves/audit/auditing/Functionality.java
@@ -1,5 +1,0 @@
-package com.involves.audit.auditing;
-
-public enum Functionality {
-	CUSTOMER, FORM;
-}
