+++ webapp/src/main/java/com/involves/audit/configuration/AppConfiguration.java
@@ -18,6 +18,9 @@ import io.prometheus.client.exporter.MetricsServlet;
 
 @Configuration
 public class AppConfiguration {
+	
+	public static String USER_NAME = "wesley.ramos";
+	public static String APPLICATION_NAME = "devepoler";
 
 	@Bean
 	public ModelMapper modelMapper() {
