+++ webapp/src/main/java/com/involves/audit/services/customer/CustomerAudit.java
@@ -1,23 +1,24 @@
-package com.involves.audit.auditing;
+package com.involves.audit.services.customer;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.event.TransactionPhase;
 import org.springframework.transaction.event.TransactionalEventListener;
 
-import com.involves.audit.services.customer.CustomerCreatedEvent;
-import com.involves.audit.services.form.FormCreatedEvent;
+import com.involves.audit.auditing.Auditing;
+import com.involves.audit.auditing.AuditingBuilder;
+import com.involves.audit.auditing.AuditingService;
 
 @Service
-public class AuditingListener {
+public class CustomerAudit {
 	
 	private AuditingService auditingService;
 	
 	@Autowired
-	public AuditingListener(AuditingService auditingService) {
+	public CustomerAudit(AuditingService auditingService) {
 		this.auditingService = auditingService;
 	}
-	
+
 	@TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
     public void processCustomerCreatedEvent(CustomerCreatedEvent customerEvent) {
 		
@@ -29,20 +30,6 @@ public class AuditingListener {
 			.withField("profileId", customerEvent.getCustomer().getProfileId())
 		.build();
 			
-		auditingService.audit(Functionality.CUSTOMER, auditing);
-    }
-	
-	
-	@TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
-    public void processFormCreatedEvent(FormCreatedEvent formEvent) {
-		
-		Auditing  auditing = AuditingBuilder.aAuditing()
-			.withField("actor", formEvent.getUsername())
-			.withField("id", formEvent.getForm().getId())
-			.withField("name", formEvent.getForm().getName())
-			.withField("goal", formEvent.getForm().getGoal())
-		.build();
-			
-		auditingService.audit(Functionality.FORM, auditing);
+		auditingService.audit("CUSTOMER_SAVE", auditing);
     }
 }
